Add SSM-Agent version on each request to the StationService
Added the agent's name and its version on the User-Agent field of each request directed to the StationService in order to have this information on the StationService and use it if needed.
E.g. Customer might run different versions of the SSM-Agent. If we release a new ssm-agent it might be interesting/important to know what version distribution we have and how many outdated callers.