Add tracer interface for tracing configurepackage runs
The interface should keep track of different steps performed while
configurepackage run. It should be a wrapper around basic output logging
and collecting output for PluginOutput. Also in the end the traces
get reported to PackageService. There are two methods
`ToPackageServiceTrace` and `ToPluginOutput` which apply the
transformation to the required output format.

Collecting the overall `Status` and `Exitcode` as required for
`PluginOutput` and `PackageService` is not part of the tracer and should
be done independently of it. Therefore it does not implement the
`PluginOutput` interface.

Basically the tracer has two modes for submitting traces:

* Starting a Trace and ending it explicitly later to keep track of the
  duration.
* Constructing a Trace manually and submitting it as a one time thing
  without a duration.

A trace can track the ExitCode and the Error. If no ExitCode is set but
an Error the ExitCode 1 will be assumed.

cr https://code.amazon.com/reviews/CR-374137