[Major Change] Changed RPM installer to not start the service by default after initial installation
Starting in the next version of SSM Agent, to align with the guidelines for RPM-based distros, the RPM intaller won't start the service by default after initial RPM installation. It will start automatically after RPM update only if the service was running previously. The behavior of the Debian and MSI installer is unchanged

cr https://code.amazon.com/reviews/CR-658367