Removed the error message of "Package is incomplete" from the output.
Before we refactor the logging messages, instead of confusing customers with an error message and then a success, we only keep this error in traces.
In case this error persists, it will be caught in traces.

sim: https://issues.amazon.com/issues/birdwatcher-597

cr https://code.amazon.com/reviews/CR-997660