Log 'Unexpected 'TimeoutSeconds' value %v received' as INFO
Switch the `Unexpected 'TimeoutSeconds' value %v received` message to INFO level, to match the log level of the surrounding messages.

The other 2 log messages in the surrounding code are already INFO level logs, so it seems odd that that line is set to ERROR level. As a default's being set, this doesn't really seem like an error, and just causes lots of log entries in the error log, that make finding true errors harder to spot.