Allow to use the package folder name to get information on the package
This will allow to use the package folder name to retrieve information on the package as well as the package arn. That means that if the name used is already normalized (birdwatcher package folder name) will be used as it is instead of normalize it.

cr https://code.amazon.com/reviews/CR-1435111