Override String() for the InstallState enum
Added the String() function for the InstallState enum in order to print the value of the enum when the %s is used in the log. Eg. Currently "state: %s" prints "state:%!s(localpackages.InstallState=7)" with this change it will print  "state: Installing" . This would help to debug the collected traces

cr https://code.amazon.com/reviews/CR-1975974