Avoid metadata lookup on on-premises instances
Metadata lookups for instanceType and availabilityZone will fail on
on-premises instances, resulting in empty strings. However, each time
this is attempted takes about 100-120 seconds, and the result is not
recognized as a cached value.

If we know it is an on-premises instance, then we can instantly return
a dummy value for instanceType and availabilityZone, speeding up the
overhead from 7-8 minutes to 1-2 seconds.

Issue: https://sim.amazon.com/issues/birdwatcher-1039

cr https://code.amazon.com/reviews/CR-2210355