Fix for long registration times under MacOS.
Macs don't have instance metadata services, so querying them will always fail. Due to some earlier issues, the usual mitigation for this in the SDK was disabled and the number of retries boosted to 10 (from an earlier 5). This change splits the configuration between MacOS and non-MacOS such that the retries are capped at 0 and the timeout mitigations are enabled for MacOS machines. There are some codepaths that still call the instance metadata service in an non-optimal fashion, but the fix for these is deferred until a more thorough design can be finalized.