Translate installationType from ConfigureAWSPackage document to Install/Update actions
 ###  Motivation
We want to introduce an installationType to specify if the update is in place or existing uninstall->install behavior

 ### Detailed Description
The actions are passed from ConfigureAWSPackage document down to the agent. With the previous "Update" action defined, it's not so obvious that "Install" action infers install or legacy update. With the new "in-place update" feature, we still want to keep the previous legacy update behavior to give customers the flexibilities.

Therefore, we will keep the "Install" and "Uninstall" actions while introducing a new parameter called "installationType". The UpdateType has two enum values that specify if the update is legacy or in place. We will parse this parameter from input, and translate to the "Install", "Update" and "Uninstall" action in "agent", which is different context from the "Install" and "Uninstall" in command actions.

Other changes that are along with this CR:
* When the package does not exist on the host when customer chooses in-place update, use the install script to execute install action (by setting isUpdateInPlace to false)
* When update script is not present when customer chooses in-place update, fail the execution while leaving the package in installed state

 ### Test Notes
bbr

cr https://code.amazon.com/reviews/CR-14772841