Generate session logs without initalizing a pty
This commit removes the need of initializing a pty and using GNU
screen for linux when generating logs. The new approach will simply run
the commands required as a one-liner script, rather than initializing
a pty and piping in the commands separately to the stdin of the pty.
This change will fix the race condition when uploading logs.

cr https://code.amazon.com/reviews/CR-20698821