enable sessions when running without root rights
This commit enables the creation of sessions when the agent runs without
root rights (as determined by geteuid) on Linux/Darwin.  Sessions are
started without root rights use the permissions that the agent has.
Sessions requested with 'Run As' fail with an appropriate error message.

This change primarily benefits developers who are testing the agent, but may
also benefit use-cases where running the agent with restricted permissions
is desirable.

The behavior in case the agent does run with root rights (the standard case
right now) is not changed.  The behavior on Windows is not changed.

cr: https://code.amazon.com/reviews/CR-62983529