[Seamless DomainJoin] Domain join using non-local username
This change adds ability to use a username existing in a different domain
to domain-join instance to local domain, if AD Trust is setup between
the domains.

If AD Trust is setup between 2 domains 'onprem-domain' and 'aws-domain', a
username from onprem-domain can be used to domain-join an instance to
aws-domain. The username from 'onprem-domain' must be given permissions in
'aws-domain' to create computer objects in 'aws-domain'.

Setup: awsSeamlessDomainUsername in SecretsManager needs to be setup as
'username@onprem-domain' instead of just 'username'

Note: AD Trust requires VPC peering between the 2 domains as mentioned
in AWS documentation.

CR: https://code.amazon.com/reviews/CR-63415639