agent/s3util: log to correct logger in headBucketTransport.RoundTrip
When the delegate responds with a status code indicating a redirect but
no `Location` response header, the log message recording the status code
override is sent to the package-wide mock logger defined in
`test_s3util.go`, causing the message to be lost. Instead, send the
message to the real logger referenced in the `headBucketTransport`
struct.

Fixes #473.