Replace mock logger with silent seelog logger
Many packages rely on `agent/log`'s `NewSilentMockLog` in situations
where a logger is required for a function call but any messages sent to
it can be discarded. This introduces a dependency on the testify test
toolkit in non-test code. Replace these calls to `NewSilentMockLog` with
requests for a new real seelog logger that simply discards any messages
sent to it.

Fixes #475.