Update domainjoin_unix_script to fix SUSE issues
There was already variables for the SUSE distribution release and service pack but the SUSEConect to enable PackageHub was hardcoded for 15.1. Newer releases have been out for years with 15.4 being recently released in 06/2022.

It's not clear to me why NetworkManager was being installed. This should not be used on SLE Server deployments. NetworkManager is only supported by SUSE for desktop workloads with SLED or the Workstation extension. All server certifications are done with wicked as the network configuration tool and using NetworkManager may invalidate them. NetworkManager is not supported by SUSE for any server workloads.

cr: https://code.amazon.com/reviews/CR-102700734