go.mod: add chainguard.dev/apko requires
Note, not updating vendor files, as the amount of updates there is
huge.

Please let me know how to proceed.

Likely splitting out / reimplementing reading and parsing a single
text file inline might be a better approach, then pulling in all of
chainguard.dev/apko which has a lot of functionality, which is unused
by the inventory integration feature.